demux: libxmp wrapper
libxmp is a library that decodes module files, such as .mod.

All decoding is done inside the demuxer, because libxmp basically turns
a filename into raw PCM data. libxmp doesn't support input I/O
abstraction either, so only local files can be played.

Since libxmp as standalone library is still experimental and API
changes are to be expected and support for it in a video player isn't
exactly a killer feature, this probably won't be merged any time soon.