demux_mkv: support CueRelativePosition
This is supposed to make seeking faster. Except that it forces you to do
dumb stuff like doing _two_ disk level seeks instead of one.

Note that all mkvmerge versions until mkvtoolnix git 0d395f9 produced
broken values for this element, and made it point to a Block instead of
a BlockGroup. We have to check for this explicitly too.

This doesn't handle preroll stuff etc. (in fact, if the current sub-
preroll hack is enabled, the new code is not used at all). We also read
CueDuration, but don't do anything with it yet.

In some places we explicitly use stream_peek() to guarantee that we can
seek back by small amounts. Add a comment to the function to document
this usage.