demux_mkv: fix some potential issues with small backwards seeks
In some cases, the demuxer seeks backwards to unread EBML tags, so that
other code can parse the properly. In theory, it could happen that a
small backwards seek requires the stream buffer to be refilled (for
example, because the preceding read caused a buffer refill and threw
away the previous contents.

So insert some strategical stream_peek() calls, which will transparently
memmove() the buffer contents if needed.

(Also, get rid of the pos variable in ebml_resync_cluster. This is an
unrelated cosmetic change, I guess.)