wayland/shm: use wl_viewport
Lets the compositor scale the video. This relies on the compositors scaling
mechanism and efficiency. But should by far more efficient than ffmpeg scaling
which burns CPU.

In my tests mpv went from 66 % cpu load to 9 %. But the scaling wasn't as nice
as ffmpegs.

Falls back to software scaling if the compositor dosen't advertise the
wl_scaler interface.